requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
libcst==0.3.23
mypy-extensions==0.4.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
